feat: add collections blank page (#479)
